Created a model for animation using fires as a test
The new animation model uses a coroutine running on a script on the
TileMap of the level. This sets a universal frame index for all flames
to use in rendering their texture. This makes all of the flames dance
in unison. I figured this is what we want to go with for the look and
feel of the city we've previously discussed.

Buildings and trucks and NPCs can all have their own coroutines for
tempo within the unified animator. Sprite sheets can be designed such
that there are columns for each frame in the animation cycle and rows
for each possible state of the animating sprite.