Created a selected state indicator for trucks
As a result found a bug with selecting and sending trucks:
the vision gameobject was catching clicks/taps and causing trucks
near a destination to become selected. This has been fixed
by creating an interactable layer to use in the physics raycast.
Currently the body of the truck and the map are the only objects
that are assigned to that layer